+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added documentation on options page about how to change the keyboard shortcut to open the extension.
 - Fixed bug where help text would show in the popup even if a project shortcut had been added.
+- Fixed bug where renaming the active project would deselect it and require you to reselect it in the popup.
 
 0.0.5